// apps/api/src/crm/search-types/organization-search.entity.ts
export interface ICrmOrganizationSearchEntity {
	nameRu: string;
	nameEn: string;
	phones: string;
	emails: string;
	website: string;
	requisiteName: string;
	requisiteInn: string;
	contactsName: string;
	contactsPhones: string;
	contactsEmails: string;
	isArchive: number;
	isReserve: number;
	isVerified: number;
	userId: number;
}